Moved sub-project to root directory.
