--- jMusixMatch/src/main/java/org/jmusixmatch/config/Constants.java
@@ -1,80 +1,0 @@
-package org.jmusixmatch.config;
-
-/**
- * A class to hold the Constants values.
- *
- * @author Sachin Handiekar
- * @version 1.0
- */
-public final class Constants {
-
-    /**
-     * A constant field to denote the value of API_KEY in the KV mapping.
-     */
-    public static final String API_KEY = "apikey";
-
-    /**
-     * A constant field to denote the base API url.
-     */
-    public static final String API_URL = "http://api.musixmatch.com/ws/";
-
-    /**
-     * A constant field to denote the API version.
-     */
-    public static final String API_VERSION = "1.1";
-
-    /**
-     * A constant field to denote the value of FORMAT in the KV mapping.
-     */
-    public static final String FORMAT = "format";
-
-    /**
-     * A constant field to denote the value of F_HAS_LYRICS in the KV mapping.
-     */
-    public static final String F_HAS_LYRICS = "f_has_lyrics";
-
-    /**
-     * A constant field to denote the value of JSON in the KV mapping.
-     */
-    public static final String JSON = "json";
-
-    /**
-     * A constant field to denote the value of PAGE in the KV mapping.
-     */
-    public static final String PAGE = "page";
-
-    /**
-     * A constant field to denote the value of PAGE_SIZE in the KV mapping.
-     */
-    public static final String PAGE_SIZE = "page_size";
-
-    /**
-     * A constant field to denote the value of QUERY in the KV mapping.
-     */
-    public static final String QUERY = "q";
-
-    /**
-     * A constant field to denote the value of QUERY_ARTIST in the KV mapping.
-     */
-    public static final String QUERY_ARTIST = "q_artist";
-
-    /**
-     * A constant field to denote the value of QUERY_TRACK in the KV mapping.
-     */
-    public static final String QUERY_TRACK = "q_track";
-
-    /**
-     * A constant field to denote the value of TRACK_ID in the KV mapping.
-     */
-    public static final String TRACK_ID = "track_id";
-
-    /**
-     * A constant field to denote the value of '/' .
-     */
-    public static final String URL_DELIM = "/";
-
-    /**
-     * A constant field to denote the value of XML in the KV mapping.
-     */
-    public static final String XML = "xml";
-}
