+++ src/main/java/org/jmusixmatch/entity/lyrics/get/LyricsGetContainer.java
@@ -1,0 +1,30 @@
+package org.jmusixmatch.entity.lyrics.get;
+
+import com.google.gson.annotations.SerializedName;
+
+import org.jmusixmatch.entity.Header;
+
+public class LyricsGetContainer {
+	
+    @SerializedName("body")
+    private LyricsGetBody body;
+    
+    @SerializedName("header")
+    private Header header;
+
+    public Header getHeader() {
+        return header;
+    }
+
+    public void setHeader(Header header) {
+        this.header = header;
+    }
+
+    public LyricsGetBody getBody() {
+        return body;
+    }
+
+    public void setBody(LyricsGetBody body) {
+        this.body = body;
+    }
+}
